requirements.txt
mysql-connector-python==8.0.26
pymongo==3.12.0
requests==2.26.0
tornado==6.1
xlwt
